Refactor icons, header and pages
