+++ components/icons/BarIcon.js
@@ -1,10 +1,8 @@
 import React from 'react'
 
-const BarIcon = (props) => (
+export const BarIcon = (props) => (
   <svg {...props} xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
     <path d="M0 0h24v24H0z" fill="none" />
     <path d="M21 5V3H3v2l8 9v5H6v2h12v-2h-5v-5l8-9zM7.43 7L5.66 5h12.69l-1.78 2H7.43z" />
   </svg>
 )
-
-export default BarIcon
